data/requirements.txt
# Vehicle Troubleshooting Chatbot - Python Dependencies

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# NLP Libraries
nltk>=3.8.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Google Gemini API
google-generativeai>=0.3.0

# Speech Processing
SpeechRecognition>=3.10.0
gTTS>=2.4.0

# Web Framework (Backend API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Firebase Integration
firebase-admin>=6.3.0

# Database & Vector Storage
chromadb>=0.4.0
scipy>=1.11.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# For production deployment
gunicorn>=21.2.0  # Production WSGI server
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
